refactor: separate user models into app/models/user.py
